simplify
fix: remove template

fix: job naming

fix namingo

fix: build

fix: install

fix: add envs

set timeout

set timeout

change pool & fix test

change pool

disable test

separate container tests

add timeout to be safe

stop on failure

use host network

re-enable test

use default pool

debug

fix cs

debug

fix pipeline

debug

debug

fix test

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

split tests

split tests & enable tests

cleanup

split tests

split tests
