--- .travis.yml
@@ -1,89 +1,0 @@
-sudo: required
-
-language: php
-php: 5.6
-
-cache:
-  directories:
-   - vendor
-
-env:
-  global:
-    # for cc-test-reporter
-    - GIT_COMMIT_SHA=$TRAVIS_COMMIT
-    - GIT_BRANCH=$TRAVIS_BRANCH
-
-jobs:
-  include:
-    - stage: prepare
-      script:
-        - composer install --no-scripts
-        - ./vendor/bin/phpcs --standard=psr2 --ignore=vendor -n .
-        #- curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
-        #- chmod +x ./cc-test-reporter
-        #- ./cc-test-reporter before-build
-        - docker info
-    - stage: tests
-      name: "Base Tests"
-      script:
-        - ./vendor/bin/phpunit --testsuite base-tests --coverage-clover build/logs/clover.xml --debug
-        #- curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
-        #- chmod +x ./cc-test-reporter
-        #- ./cc-test-reporter format-coverage --output "coverage/codeclimate.1.json"
-        #- docker run --rm -e AWS_ACCESS_KEY_ID=$AWS_ECR_ACCESS_KEY_ID -e AWS_SECRET_ACCESS_KEY=$AWS_ECR_SECRET_ACCESS_KEY -v $TRAVIS_BUILD_DIR/coverage/:/data/ quay.io/keboola/aws-cli:latest s3 sync /data/ "s3://docker-runner-logs/coverage/$TRAVIS_BUILD_NUMBER"
-    - stage: tests
-      name: "Runner Tests"
-      script:
-        - ./vendor/bin/phpunit --testsuite runner-tests --coverage-clover build/logs/clover.xml --debug
-        #- curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
-        #- chmod +x ./cc-test-reporter
-        #- ./cc-test-reporter format-coverage --output "coverage/codeclimate.2.json"
-        #- docker run --rm -e AWS_ACCESS_KEY_ID=$AWS_ECR_ACCESS_KEY_ID -e AWS_SECRET_ACCESS_KEY=$AWS_ECR_SECRET_ACCESS_KEY -v $TRAVIS_BUILD_DIR/coverage/:/data/ quay.io/keboola/aws-cli:latest s3 sync /data/ "s3://docker-runner-logs/coverage/$TRAVIS_BUILD_NUMBER"
-    - stage: tests
-      name: "Executor Tests"
-      script:
-        - ./vendor/bin/phpunit --testsuite executor-tests --coverage-clover build/logs/clover.xml --debug
-        #- curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
-        #- chmod +x ./cc-test-reporter
-        #- ./cc-test-reporter format-coverage --output "coverage/codeclimate.3.json"
-        #- docker run --rm -e AWS_ACCESS_KEY_ID=$AWS_ECR_ACCESS_KEY_ID -e AWS_SECRET_ACCESS_KEY=$AWS_ECR_SECRET_ACCESS_KEY -v $TRAVIS_BUILD_DIR/coverage/:/data/ quay.io/keboola/aws-cli:latest s3 sync /data/ "s3://docker-runner-logs/coverage/$TRAVIS_BUILD_NUMBER"
-    - stage: tests
-      name: "S3 Tests"
-      script:
-        - ./vendor/bin/phpunit --testsuite s3-tests --coverage-clover build/logs/clover.xml --debug
-        #- curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
-        #- chmod +x ./cc-test-reporter
-        #- ./cc-test-reporter format-coverage --output "coverage/codeclimate.4.json"
-        #- docker run --rm -e AWS_ACCESS_KEY_ID=$AWS_ECR_ACCESS_KEY_ID -e AWS_SECRET_ACCESS_KEY=$AWS_ECR_SECRET_ACCESS_KEY -v $TRAVIS_BUILD_DIR/coverage/:/data/ quay.io/keboola/aws-cli:latest s3 sync /data/ "s3://docker-runner-logs/coverage/$TRAVIS_BUILD_NUMBER"
-    - stage: tests
-      name: "ABS Tests"
-      script:
-        - export STORAGE_API_TOKEN=$STORAGE_API_TOKEN_ABS; ./vendor/bin/phpunit --testsuite abs-tests --coverage-clover build/logs/clover.xml --debug
-        #- curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
-        #- chmod +x ./cc-test-reporter
-        #- ./cc-test-reporter format-coverage --output "coverage/codeclimate.5.json"
-        #- docker run --rm -e AWS_ACCESS_KEY_ID=$AWS_ECR_ACCESS_KEY_ID -e AWS_SECRET_ACCESS_KEY=$AWS_ECR_SECRET_ACCESS_KEY -v $TRAVIS_BUILD_DIR/coverage/:/data/ quay.io/keboola/aws-cli:latest s3 sync /data/ "s3://docker-runner-logs/coverage/$TRAVIS_BUILD_NUMBER"
-    - stage: tests
-      name: "Synapse Tests"
-      script:
-        - export STORAGE_API_TOKEN=$STORAGE_API_TOKEN_ABS; ./vendor/bin/phpunit --testsuite synapse-tests --coverage-clover build/logs/clover.xml --debug
-        #- curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
-        #- chmod +x ./cc-test-reporter
-        #- ./cc-test-reporter format-coverage --output "coverage/codeclimate.6.json"
-        #- docker run --rm -e AWS_ACCESS_KEY_ID=$AWS_ECR_ACCESS_KEY_ID -e AWS_SECRET_ACCESS_KEY=$AWS_ECR_SECRET_ACCESS_KEY -v $TRAVIS_BUILD_DIR/coverage/:/data/ quay.io/keboola/aws-cli:latest s3 sync /data/ "s3://docker-runner-logs/coverage/$TRAVIS_BUILD_NUMBER"
-    #- stage: after-tests
-    #  script:
-        #- docker run --rm -e AWS_ACCESS_KEY_ID=$AWS_ECR_ACCESS_KEY_ID -e AWS_SECRET_ACCESS_KEY=$AWS_ECR_SECRET_ACCESS_KEY -v $TRAVIS_BUILD_DIR/coverage/:/data/ quay.io/keboola/aws-cli:latest s3 sync "s3://docker-runner-logs/coverage/$TRAVIS_BUILD_NUMBER" /data/
-        #- curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
-        #- chmod +x ./cc-test-reporter
-        #- ./cc-test-reporter sum-coverage --output - --parts 6 coverage/codeclimate.*.json | ./cc-test-reporter upload-coverage --input -
-
-notifications:
-  email: false
-  slack:
-    secure: doRb1zL09mK/dGUFgXAz3TqSAkCASbYMQYrGbOtfolIiBHwDHpO8J7QDXo2OHCUvYd4bOMv2HAluooz4nYn50RskUdYDkqUTgDRLVmRwdsH3KiBdQ5awnxfbZTvXPPGKSAbbOaH4AF+ig++I9oBJqzF5pBvcp2sfOiQkIrsjKr8=
-
-branches:
-  only:
-  - master
-  - production
