+++ Tests/RunnerPart2/Runner2Test.php
@@ -1,0 +1,86 @@
+<?php
+
+namespace Keboola\DockerBundle\Tests\Runner;
+
+use Keboola\DockerBundle\Docker\Component;
+use Keboola\DockerBundle\Docker\JobDefinition;
+use Keboola\DockerBundle\Docker\Runner\UsageFile\NullUsageFile;
+use Keboola\DockerBundle\Exception\UserException;
+use Keboola\DockerBundle\Tests\BaseRunnerTest;
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Process\Process;
+
+class Runner2Test extends BaseRunnerTest
+{
+    public function setUp()
+    {
+        parent::setUp();
+    }
+
+    public function testPermissionsFailedWithoutContainerRootUserFeature()
+    {
+        $componentData = [
+            'id' => 'keboola.docker-demo-sync',
+            'data' => [
+                'definition' => [
+                    'type' => 'aws-ecr',
+                    'uri' => '147946154733.dkr.ecr.us-east-1.amazonaws.com/developer-portal-v2/keboola.python-transformation',
+                ],
+            ],
+        ];
+
+        $config = [
+            'storage' => [
+                'output' => [
+                    'tables' => [
+                        [
+                            'source' => 'mytable.csv.gz',
+                            'destination' => 'in.c-runner-test.mytable',
+                            'columns' => ['col1'],
+                        ],
+                    ],
+                ],
+            ],
+            'parameters' => [
+                'script' => [
+                    'from subprocess import call',
+                    'import os',
+                    'os.makedirs("/data/out/tables/mytable.csv.gz")',
+                    'call(["chmod", "000", "/data/out/tables/mytable.csv.gz"])',
+                    'with open("/data/out/tables/mytable.csv.gz/part1", "w") as file:',
+                    '   file.write("value1")',
+                ],
+            ],
+        ];
+
+        $runner = $this->getRunner();
+        self::expectException(UserException::class);
+        // touch: cannot touch '/data/out/tables/mytable.csv.gz/part1': Permission denied
+        self::expectExceptionMessageRegExp('/Permission denied/');
+        $runner->run(
+            $this->prepareJobDefinitions(
+                $componentData,
+                'runner-configuration',
+                $config,
+                []
+            ),
+            'run',
+            'run',
+            '1234567',
+            new NullUsageFile()
+        );
+    }
+
+    /**
+     * @param array $componentData
+     * @param $configId
+     * @param array $configData
+     * @param array $state
+     * @return JobDefinition[]
+     */
+    protected function prepareJobDefinitions(array $componentData, $configId, array $configData, array $state)
+    {
+        $jobDefinition = new JobDefinition($configData, new Component($componentData), $configId, 'v123', $state);
+        return [$jobDefinition];
+    }
+}
