--- azure-pipelines-env-template.yml
@@ -1,17 +1,0 @@
-env:
-  AWS_ECR_ACCESS_KEY_ID: $(AWS_ECR_ACCESS_KEY_ID)
-  AWS_ECR_REGISTRY_ACCOUNT_ID: $(AWS_ECR_REGISTRY_ACCOUNT_ID)
-  AWS_ECR_REGISTRY_REGION: $(AWS_ECR_REGISTRY_REGION)
-  AWS_ECR_REGISTRY_URI: $(AWS_ECR_REGISTRY_URI)
-  AWS_ECR_SECRET_ACCESS_KEY: $(AWS_ECR_SECRET_ACCESS_KEY)
-  AWS_KMS_TEST_KEY: $(AWS_KMS_TEST_KEY)
-  DOCKERHUB_PRIVATE_PASSWORD: $(DOCKERHUB_PRIVATE_PASSWORD)
-  DOCKERHUB_PRIVATE_SERVER: $(DOCKERHUB_PRIVATE_SERVER)
-  DOCKERHUB_PRIVATE_USERNAME: $(DOCKERHUB_PRIVATE_USERNAME)
-  GIT_PRIVATE_PASSWORD: $(GIT_PRIVATE_PASSWORD)
-  GIT_PRIVATE_USERNAME: $(GIT_PRIVATE_USERNAME)
-  QUAYIO_PRIVATE_PASSWORD: $(QUAYIO_PRIVATE_PASSWORD)
-  QUAYIO_PRIVATE_USERNAME: $(QUAYIO_PRIVATE_USERNAME)
-  STORAGE_API_TOKEN: $(STORAGE_API_TOKEN)
-  STORAGE_API_URL: $(STORAGE_API_URL)
-  STORAGE_API_TOKEN_ABS: $(STORAGE_API_TOKEN_ABS)
