--- build.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<project name="api-tests" default="build">
-
-    <property name="storage-api-token" value="your_token"/>
-
-    <target name="build" depends="prepare,api-tests-ci"/>
-    <target name="build-full" depends="prepare,api-tests-ci-full"/>
-
-    <target name="clean" description="Cleanup build artifacts">
-        <delete dir="${basedir}/build/results/api"/>
-        <delete dir="${basedir}/build/results/code-browser"/>
-        <delete dir="${basedir}/build/results/coverage"/>
-        <delete dir="${basedir}/build/results/logs"/>
-        <delete dir="${basedir}/build/results/pdepend"/>
-        <delete dir="${basedir}/build/results/apigen"/>
-    </target>
-
-    <target name="prepare" depends="clean"
-            description="Prepare for build">
-        <mkdir dir="${basedir}/build/results/api"/>
-        <mkdir dir="${basedir}/build/results/code-browser"/>
-        <mkdir dir="${basedir}/build/results/coverage"/>
-        <mkdir dir="${basedir}/build/results/logs"/>
-        <mkdir dir="${basedir}/build/results/pdepend"/>
-        <mkdir dir="${basedir}/build/results/apigen"/>
-        <exec executable="/usr/local/bin/composer" failonerror="true">
-            <arg value="install"/>
-            <env key="KEBOOLA_SYRUP_CONFIGS_BUCKET" value="${syrup-configs-bucket}"/>
-            <env key="AWS_REGION" value="${aws-region}"/>
-        </exec>
-    </target>
-
-    <target name="create-index" depends="prepare" description="Create index">
-        <exec executable="./vendor/keboola/syrup/app/console" failonerror="true">
-            <arg value="syrup:create-index" />
-        </exec>
-    </target>
-
-    <target name="api-tests-ci" depends="prepare" description="Run unit tests with PHPUnit">
-        <exec executable="${basedir}/vendor/bin/phpunit" failonerror="true">
-            <env key="STORAGE_API_URL" value="${storage-api-url}"/>
-            <env key="STORAGE_API_TOKEN" value="${storage-api-token}"/>
-            <env key="GIT_PRIVATE_USERNAME" value="${git-private-username}"/>
-            <env key="GIT_PRIVATE_PASSWORD" value="${git-private-password}"/>
-            <env key="AWS_ECR_ACCESS_KEY_ID" value="${aws-ecr-access-key-id}"/>
-            <env key="AWS_ECR_SECRET_ACCESS_KEY" value="${aws-ecr-secret-access-key}"/>
-            <env key="AWS_ECR_REGISTRY_URI" value="${aws-ecr-registry-uri}"/>
-            <env key="AWS_ECR_REGISTRY_REGION" value="${aws-ecr-registry-region}"/>
-            <env key="AWS_KMS_TEST_KEY" value="${aws-kms-test-key}"/>
-            <arg value="--configuration"/>
-            <arg path="phpunit.xml.dist"/>
-            <arg value="--stop-on-error"/>
-            <arg value="--stop-on-failure"/>
-        </exec>
-    </target>
-
-    <target name="api-tests-ci-full" depends="create-index" description="Run unit tests with PHPUnit">
-        <exec executable="${basedir}/vendor/bin/phpunit" failonerror="true">
-            <env key="STORAGE_API_URL" value="${storage-api-url}"/>
-            <env key="STORAGE_API_TOKEN" value="${storage-api-token}"/>
-            <env key="GIT_PRIVATE_USERNAME" value="${git-private-username}"/>
-            <env key="GIT_PRIVATE_PASSWORD" value="${git-private-password}"/>
-            <env key="AWS_ECR_ACCESS_KEY_ID" value="${aws-ecr-access-key-id}"/>
-            <env key="AWS_ECR_SECRET_ACCESS_KEY" value="${aws-ecr-secret-access-key}"/>
-            <env key="AWS_ECR_REGISTRY_URI" value="${aws-ecr-registry-uri}"/>
-            <env key="AWS_ECR_REGISTRY_REGION" value="${aws-ecr-registry-region}"/>
-            <env key="AWS_KMS_TEST_KEY" value="${aws-kms-test-key}"/>
-            <env key="AWS_ECR_REGISTRY_ACCOUNT_ID" value="${aws-ecr-registry-account-id}"/>
-            <arg value="--configuration"/>
-            <arg path="phpunit-full.xml.dist"/>
-            <arg value="--stop-on-error"/>
-            <arg value="--stop-on-failure"/>
-        </exec>
-    </target>
-
-</project>
