+++ docker-compose.yml
@@ -6,7 +6,7 @@ services:
     build:
       context: .
       dockerfile: DockerfileTests
-    image: keboola/docker-runner
+    image: keboola/docker-runner-tests
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
       - /tmp/:/tmp/
@@ -22,7 +22,15 @@ services:
       - DOCKERHUB_PRIVATE_USERNAME
       - GIT_PRIVATE_PASSWORD
       - GIT_PRIVATE_USERNAME
-      - QUAYIO_PRIVATE_PASSWORD
-      - QUAYIO_PRIVATE_USERNAME
       - STORAGE_API_TOKEN
       - STORAGE_API_URL
+      - STORAGE_API_URL_SYNAPSE
+      - STORAGE_API_TOKEN_SYNAPSE
+      - STORAGE_API_TOKEN_MASTER
+      - RUNNER_COMMAND_TO_GET_HOST_IP
+      - STORAGE_API_TOKEN_READ_ONLY
+      - RUN_SYNAPSE_TESTS
+
+  host-tests: 
+    <<: *tests
+    network_mode: host
